+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * [Controller] Add `current_path` and `current_url` to generate a connection's path and url
   * [Channel] Add ability to configure channel event logging with `:log_join` and `:log_handle_in` options
   * [Channel] Warn on unhandled `handle_info/2` messages
+  * [Router] document `match` support for matching on any http method with the special `:*` argument
 
 * JavaScript client enhancements
   * Add ability to pass `encode` and `decode` functions to socket constructor for custom encoding and decoding of outgoing and incoming messages.